# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Poll API
  version: 1.0.0
  description: API for user registration, authentication and polls.
servers:
  - url: http://localhost:3000
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Consent:
      type: object
      properties:
        tosAcceptedAt:
          type: string
          format: date-time
        tosVersion:
          type: string
        privacyAcceptedAt:
          type: string
          format: date-time
        privacyVersion:
          type: string
      required:
        - tosAcceptedAt
        - tosVersion
        - privacyAcceptedAt
        - privacyVersion
    UserPublic:
      type: object
      properties:
        id:
          type: string
          example: "256375c5-7f8c-4661-a83e-a5838b363196"
        username:
          type: string
          example: "test1"
        createdAt:
          type: string
          format: date-time
        consent:
          $ref: "#/components/schemas/Consent"
      required: [id, username, createdAt, consent]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          example: "test1"
        password:
          type: string
          example: "supersecret123"
        tosAccepted:
          type: boolean
          example: true
      required: [username, password, tosAccepted]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required: [ok]
...
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users/me:
    delete:
      tags: [Users]
      summary: Delete my user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...